+++ test/cases/config/validate-config.test.js
@@ -96,7 +96,12 @@ describe('测试: 验证配置 (生成临时的核心代码引用文件，返回
             expect(typeof kootConfig.dist).toBe('string');
             expect(typeof kootConfig.cookiesToStore !== 'undefined').toBe(true);
             expect(typeof kootConfig.i18n !== 'undefined').toBe(true);
-            expect(typeof kootConfig.serviceWorker !== 'undefined').toBe(true);
+            expect(typeof kootConfig.pwa).toBe('undefined');
+            expect(
+                ['boolean', 'object'].some(
+                    t => typeof kootConfig.serviceWorker === t
+                ) && !Array.isArray(kootConfig.serviceWorker)
+            ).toBe(true);
             expect(
                 typeof kootConfig.aliases === 'object' &&
                     !Array.isArray(kootConfig.aliases)