+++ CHANGELOG.md
@@ -13,7 +13,12 @@
         -   特殊 JS 入口 _CLIENT_RUN_FIRST_ 现默认引入 `regenerator-runtime/runtime`
             -   现在 _Koot.js_ 项目默认支持在浏览器环境中使用 _async_ / _await_ 开发
     -   默认的 _Service Worker_ 文件现在使用 _Workbox_ 生成
+        -   如果原有项目中有使用自定的 _Service Worker_ 模板，升级后需要更新该模板。详情请参见文档 [Service Worker](https://koot.js.org/#/pwa)
     -   现在会忽略 _Babel_ 插件 `@babel/plugin-transform-regenerator`
+    -   更新依赖包 **major** 版本号
+        -   `chalk` -> _^3.0.0_
+        -   `ora` -> _^4.0.3_
+        -   `webpack` -> _^5.0.0_
 -   **新特性**
     -   **新配置项** `distClientAssetsDirName` - 设定客户端打包结果中静态资源存放路径的目录名。详情请参见文档 [配置/distClientAssetsDirName](https://koot.js.org/#/config?id=distClientAssetsDirName) ([#181](https://github.com/cmux/koot/issues/181))
     -   **新工具函数** `koot/utils/client-get-styles` - 获取当前全局 CSS 和所有组件 CSS。具体用法请参见文档 [全局与工具函数/工具函数/客户端](https://koot.js.org/#/utilities?id=客户端) ([#185](https://github.com/cmux/koot/issues/185))
@@ -21,15 +26,21 @@
     -   [服务器] 优化服务器代码的文件尺寸 ([#172](https://github.com/cmux/koot/issues/172), [#186](https://github.com/cmux/koot/issues/186))
     -   [SSR & 客户端渲染 & SPA 模板生成] 所有注入/插入的 `<script>` 标签现在均会新增 `[data-koot-entry]` 属性
     -   现在开发环境下会默认启用 _Service Worker_
--   更新依赖包 major 版本号
-    -   `chalk` -> _^3.0.0_
+
+**koot-boilerplate**
+
+-   _Webpack_ 配置
+    -   图片资源改用 `url-loader` 处理
+        -   大于 **2KB** 的文件会继续使用 `file-loader` 处理
+    -   大于 **5KB** 的 _SVG_ 的文件会改用 `file-loader` 处理
 
 **koot-cli**
 
 -   错误修正
     -   新建项目时的“覆盖”模式会删除原有文件的问题 ([#180](https://github.com/cmux/koot/issues/180))
--   更新依赖包 major 版本号
+-   更新依赖包 **major** 版本号
     -   `chalk` -> _^3.0.0_
+    -   `ora` -> _^4.0.3_
 
 ## [0.11.15] - 2019-10-15